requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr!=2.1.0,>=2.0.0 # Apache-2.0
eventlet!=0.18.3,!=0.20.1,>=0.18.2 # MIT
oslo.config>=5.2.0 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.concurrency>=3.26.0 # Apache-2.0
oslo.policy>=1.35.0 # Apache-2.0
oslo.messaging>=6.3.0 # Apache-2.0
oslo.db>=4.27.0 # Apache-2.0
paramiko>=2.0.0  # LGPLv2.1+
SQLAlchemy!=1.1.5,!=1.1.6,!=1.1.7,!=1.1.8,>=1.0.10 # MIT
keystonemiddleware>=4.17.0 # Apache-2.0
autobahn>=0.10.1  # MIT License
python-neutronclient>=6.7.0 # Apache-2.0
python-designateclient>=2.11.0 # Apache-2.0
pecan!=1.0.2,!=1.0.3,!=1.0.4,!=1.2,>=1.0.0 # BSD
PyMySQL>=0.7.6  # MIT License
osprofiler>=1.5.0 # Apache-2.0
WSME>=0.8  # MIT
